after crud endpoints completion

+++ app/controllers/api/users_controller.rb
@@ -7,17 +7,20 @@ class Api::UsersController < ApplicationController
   end
 
   def create
-    user = User.new(user_params)
-    if user.save
-      render json:{data:user,message:"User successfully added",status:200}
+    if params[:user].present?
+      user = User.new(user_params)
+      if user.save
+        render json:{data:user,message:"User successfully added",status:200}
+      else
+        render json:{errors:user.errors.full_messages,status:400}
+      end
     else
-      render json:{errors:user.errors.full_messages,status:400}
+      render json:{message: "User details required"}
     end
   end
 
   def show
     user = User.find_by(id:params[:id])
-    binding.pry
     if user.present?
       render json:{data:user,status:200}
     else
@@ -26,20 +29,28 @@ class Api::UsersController < ApplicationController
   end
 
   def update
-    user = User.find(params[:id])
-    if user.update(user_params)
-      render json:{data:user,message:"User successfully updated",status:200}
+    user = User.find_by(id:params[:id])
+    if user.present?
+      if user.update(user_params)
+        render json:{data:user,message:"User successfully updated",status:200}
+      else
+        render json:{errors:user.errors.full_messages,status:400}
+      end
     else
-      render json:{errors:user.errors.full_messages,status:400}
+      render json:{message: "User not found",status:400}
     end
   end
 
   def destroy
-    user = User.find(params[:id])
-    if user.destroy
-      render json:{message: 'User successfully deleted',status:200}
-    else
-      render json:{errors: "Failed to delete user",status:400}
+    user = User.find_by(id:params[:id])
+    if user.present?
+      if user.destroy
+        render json:{}
+      else
+        render json:{errors: "Failed to delete user",status:400}
+      end
+    else 
+      render json:{message: "User not found",status:400}
     end
   end
